+++ src/plugins/storage-worker/workers/webpack.config.js
@@ -30,7 +30,7 @@ module.exports = {
         cacheDirectory: path.resolve(
             projectRootPath,
             'test_tmp',
-            'webpack-cache'
+            'webpack-cache-worker'
         ),
     },
     devtool: 'source-map',
@@ -47,7 +47,7 @@ module.exports = {
                 use: {
                     loader: 'babel-loader',
                     options: babelConfig
-                }
+                },
             }
         ],
     },